+++ pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>me.roinujnosde</groupId>
     <artifactId>TitansBattle</artifactId>
-    <version>6.0.2.5</version>
+    <version>6.0.2.6</version>
     <packaging>jar</packaging>
     <name>TitansBattle</name>
     <properties>
