+++ src/main/java/me/roinujnosde/titansbattle/commands/TBCommands.java
@@ -6,7 +6,6 @@ import co.aikar.commands.annotation.*;
 import co.aikar.commands.bukkit.contexts.OnlinePlayer;
 import me.roinujnosde.titansbattle.BaseGameConfiguration;
 import me.roinujnosde.titansbattle.TitansBattle;
-import me.roinujnosde.titansbattle.challenges.ArenaConfiguration;
 import me.roinujnosde.titansbattle.dao.ConfigurationDao;
 import me.roinujnosde.titansbattle.exceptions.CommandNotSupportedException;
 import me.roinujnosde.titansbattle.games.Game;
@@ -483,20 +482,36 @@ public class TBCommands extends BaseCommand {
                 new SimpleDateFormat(dateFormat).format(date), name, group, members));
     }
 
+    @CommandAlias("%watch|watch")
     @Subcommand("%watch|watch")
-    @CommandPermission("titansbattle.challenge.watch")
-    @CommandCompletion("@arenas:in_use")
+    @CommandPermission("titansbattle.watch")
     @Description("{@@command.description.watch}")
-    public void watch(Player sender, Game game, @Optional ArenaConfiguration arena) {
+    public void watch(Player sender, Game game) {
         BaseGameConfiguration config;
-        if (arena == null && game == null) {
-            sender.sendMessage(plugin.getLang("challenge.not-starting-or-started"));
+        if (game == null) {
+            sender.sendMessage(plugin.getLang("not-starting-or-started"));
             return;
         }
-        config = (arena == null) ? game.getConfig() : arena;
 
         Location watchroom = config.getWatchroom();
         sender.teleport(watchroom);
         SoundUtils.playSound(SoundUtils.Type.WATCH, plugin.getConfig(), sender);
     }
+
+    // @Subcommand("%watch|watch")
+    // @CommandPermission("titansbattle.challenge.watch")
+    // @CommandCompletion("@arenas:in_use")
+    // @Description("{@@command.description.watch}")
+    // public void watch(Player sender, Game game, @Optional ArenaConfiguration arena) {
+    //     BaseGameConfiguration config;
+    //     if (arena == null && game == null) {
+    //         sender.sendMessage(plugin.getLang("challenge.not-starting-or-started"));
+    //         return;
+    //     }
+    //     config = (arena == null) ? game.getConfig() : arena;
+
+    //     Location watchroom = config.getWatchroom();
+    //     sender.teleport(watchroom);
+    //     SoundUtils.playSound(SoundUtils.Type.WATCH, plugin.getConfig(), sender);
+    // }
 }
